+++ lib/zaikio/hub/person.rb
@@ -5,7 +5,9 @@ module Zaikio
       attributes :updated_at, :created_at, :first_name, :name, :full_name, :email,
                  :pronoun, :locale, :country_code, :currency, :unit_system, :connected,
                  :test_account_owner_id, :time_zone, :email_confirmed,
-                 :two_factor_authentication_enabled, :avatar_url, :subscription
+                 :two_factor_authentication_enabled, :avatar_url, :subscription,
+                 :granted_oauth_scopes, :requested_oauth_scopes,
+                 :requested_oauth_scopes_waiting_for_approval
     end
   end
 end